% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_point_selection.R
\name{interactive_point_selection}
\alias{interactive_point_selection}
\title{Interactive selection of cells from a ggplot (Seurat spatial plot)}
\usage{
interactive_point_selection(ggplot_object)
}
\arguments{
\item{ggplot_object}{A ggplot object (e.g. from ImageDimPlot)}
}
\value{
A data.frame of selected points (x, y, cell, etc.)
}
\description{
Interactive selection of cells from a ggplot (Seurat spatial plot)
}
