% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_spatial_selection.R
\name{run_spatial_selection}
\alias{run_spatial_selection}
\title{Manual spatial cell selection within a Seurat object}
\usage{
run_spatial_selection(
  seurat_obj,
  meta_id = NULL,
  fovs,
  images = NULL,
  metadata_col = "ManualSelection",
  positive_label = "Selected",
  negative_label = "Not_Selected",
  incremental = FALSE,
  point_size = 3,
  invert_coordinates = FALSE,
  cols = NULL
)
}
\arguments{
\item{seurat_obj}{A Seurat object containing spatial data.}

\item{meta_id}{Name of the metadata column identifying the FOVs (e.g. "fov", "Region", "sample_id")}

\item{fovs}{Names of the images if `meta_id` is not used (e.g. in Xenium objects)
or character vector of FOV identifiers to analyze (must match the values in `meta_id`)}

\item{images}{Optional character vector of image names (same length as `fovs`).
If NULL, the function will automatically attempt to use `fovs` as image names.}

\item{metadata_col}{Name of the metadata column to store selection results (default "ManualSelection").}

\item{positive_label}{Label for selected cells (default "Selected").}

\item{negative_label}{Label for unselected cells (default "Not_Selected").}

\item{incremental}{Logical. If TRUE, keep existing labels in `metadata_col` and add new `positive_label`
instead of resetting the column (default FALSE).}

\item{point_size}{Numeric. Size of points in the interactive selection plot (default: 3).}

\item{invert_coordinates}{Logical. If TRUE, swaps x/y coordinates in interactive selection (default: FALSE).}

\item{cols}{Optional color vector for idents passed to ImageDimPlot.}
}
\value{
The modified Seurat object with updated metadata column.
}
\description{
Manual spatial cell selection within a Seurat object
}
